requirements.txt
beautifulsoup4==4.12.2
selenium==4.12.0
webdriver-manager==4.0.0
pandas==2.1.1
tiktoken==0.5.1
python-dotenv==1.0.0
openai==0.28.0
faiss-cpu==1.7.4
langchain==0.0.300